+++ ui/pages/Search/components/filters/ProjectFamiliesFilter.jsx
@@ -9,10 +9,10 @@ import {
   getProjectDatasetTypes,
 } from 'redux/selectors'
 import { Multiselect, ButtonRadioGroup } from 'shared/components/form/Inputs'
-import { ProjectFilter } from 'shared/components/panel/search/ProjectsField'
 import { SOLVED_FAMILY_STATUS_OPTIONS } from 'shared/utils/constants'
 import { getProjectFamilies, getSearchContextIsLoading, getFamilyOptions, getAnalysisGroupOptions } from '../../selectors'
 import { loadProjectFamiliesContext } from '../../reducers'
+import { ProjectFilter } from './ProjectsField'
 
 const ALL = 'ALL'
 const UNSOLVED = 'UNSOLVED'
